requirements/base.txt
Markdown==2.4
django-filter==0.7
djangorestframework==2.3.13
wsgiref==0.1.2
git+https://github.com/django-nonrel/django@nonrel-1.6
git+https://github.com/django-nonrel/djangotoolbox
git+https://github.com/django-nonrel/mongodb-engine
